example/raspberry/requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.41
humanfriendly==10.0
idna==3.10
isodate==0.6.1
Jinja2==3.1.4
kiwisolver==1.4.7
lapx==0.5.10.post1
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
onnxruntime==1.19.2
onvif2-zeep==0.3.4
opencv-python==4.10.0.84
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.3
protobuf==5.28.1
psutil==6.0.0
py-cpuinfo==9.0.0
pyee==12.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-ffmpeg==2.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
scipy==1.14.1
seaborn==0.13.2
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sympy==1.13.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.95
ultralytics-thop==2.0.6
urllib3==2.2.3
WSDiscovery==2.0.0
zeep==4.2.1
